refactor(fury): decompose arcade-cabinet page into act-based subcomponents
The 670-line monolith becomes a 54-line orchestrator plus nine focused
subcomponents and a data module. The arcade cabinet metaphor maps onto
the component split: FuryBackground (horizon + perspective grid +
scanlines), FuryMarquee (score + title + hi-score with live ticker),
FuryCoin (insert-coin CTA), FuryHud (stats tiles), FuryOrigin (Act I),
FuryStage (single CRT-framed feature row), FurySystems (Act II — maps
features through FuryStage), FuryCredits (Act III), FuryFooter.

Keyframes (fury-bob, coin-blink, coin-pulse, feed-pulse, noise-shift,
sys-in) stay colocated with the component that owns the animated element.
The 3D perspective grid, mix-blend-mode scanlines overlay, and mask-image
vignette — all thesis-defining — remain in scoped CSS.

The `.fury-act + .fury-act` sibling selector that drew a dashed rule
between acts used to work within a single component; now the children
are separate components so I lifted it to Fury.vue with :deep() so the
parent still sees child roots as direct siblings.

Live score ticker state (setInterval in onMounted) moves into
FuryMarquee where it is used.

No visual change: 'bun run test:visual -- --grep fury' green on
mobile/tablet/desktop.
